FYP-63 refactoring responses [alethia-service]
