+++ hive-shared/src/main/java/com/fyp/hiveshared/api/responses/ResponseBody.java
@@ -1,23 +1,30 @@
-package com.fyp.hiveshared.api.response;
+package com.fyp.hiveshared.api.responses;
 
 import org.springframework.http.HttpStatus;
 
-public class ApiResponse {
+public class ResponseBody {
     private String message;
     private int status;
     private HttpStatus method;
     private String service;
     private String timestamp;
 
-    public ApiResponse() {
+    public ResponseBody() {
 
     }
 
-    public ApiResponse(String message, int status, HttpStatus method, String service, String timestamp) {
+    public ResponseBody(String message, int status, HttpStatus method, String service, String timestamp) {
         this.message = message;
         this.status = status;
         this.method = method;
         this.service = service;
+        this.timestamp = timestamp;
+    }
+
+    public ResponseBody(String message, int status, HttpStatus method, String timestamp) {
+        this.message = message;
+        this.status = status;
+        this.method = method;
         this.timestamp = timestamp;
     }
