+++ hive-shared/src/main/java/com/fyp/hiveshared/api/responses/ResponseHandler.java
@@ -1,4 +1,4 @@
-package com.fyp.hiveshared.api.response;
+package com.fyp.hiveshared.api.responses;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +9,7 @@ import java.time.ZonedDateTime;
 public class ResponseHandler {
 
     public static ResponseEntity<Object> serviceResponse(String message, HttpStatus status, String service) {
-        return new ResponseEntity<>(new ApiResponse(
+        return new ResponseEntity<>(new ResponseBody(
                 message,
                 status.value(),
                 status,
