+++ hive-shared/src/main/java/com/fyp/hiveshared/api/responses/excpetion/ApiExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.fyp.hiveshared.api.exception;
+package com.fyp.hiveshared.api.responses.excpetion;
 
+import com.fyp.hiveshared.api.responses.ResponseBody;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,24 +16,27 @@ import java.time.ZonedDateTime;
 @ControllerAdvice
 public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 
+    private static final String ALETHIA_SERVICE = "alethia-service";
+
     @Override
     protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers,
                                                                   HttpStatus status, WebRequest request) {
-        return new ResponseEntity<>(new ApiException(
+        return new ResponseEntity<>(new ResponseBody(
                 "Validation failed for some of the arguments. Make sure that all arguments are correct",
                 status.value(),
                 status,
-                ZonedDateTime.now(ZoneId.of("Z"))
+                ZonedDateTime.now(ZoneId.of("Z")).toString()
         ), status);
     }
 
     @ExceptionHandler({IdpalRequestException.class})
     public ResponseEntity<Object> handleIdpalRequestException(IdpalRequestException e) {
-        return new ResponseEntity<>(new ApiException(
+        return new ResponseEntity<>(new ResponseBody(
                 e.getMessage(),
                 HttpStatus.SERVICE_UNAVAILABLE.value(),
                 HttpStatus.SERVICE_UNAVAILABLE,
-                ZonedDateTime.now(ZoneId.of("Z"))
+                ALETHIA_SERVICE,
+                ZonedDateTime.now(ZoneId.of("Z")).toString()
         ), HttpStatus.SERVICE_UNAVAILABLE);
     }
 }
