+++ hive-shared/src/main/java/com/fyp/hiveshared/api/responses/excpetion/IdpalRequestException.java
@@ -1,4 +1,4 @@
-package com.fyp.hiveshared.api.exception;
+package com.fyp.hiveshared.api.responses.excpetion;
 
 public class IdpalRequestException extends RuntimeException{
     public IdpalRequestException(String message) {
